refactor: move test fixtures to dedicated __fixtures__ directory
