+++ apps/client/src/__fixtures__/seedDispositifs.ts
